Returns the sum of two positive numbers of indefinite sizes as string
Try and except is used when both numbers aren't of equal number of digits. Hence, upon our traditional way of addition, we get an IndexError when we have summed all the digits of the smallest number and we are left only with the remaining digits of the longest number and the carryout.