Ceasar cipher encryption
Here, the program is written in such a way that the function takes a text to be encrypted and a key. The key is used to rotate the original alphabet by shifting it to the right depending on the key value (This is to form the 'ceasar alphabet'. Each character is gotten from the text to be encrypted and its position in the traditional alphabets gotten. This index now is used on the ceasar alphabet and added to the result. 
String slicing was equally used to rotate the alphabet